Github fetcher for open issues with labels
The code expects a repo name. However, I decided to use a token key to authenticate not username or password so this is something you'd have to use on your own because mine expires shortly. There are some areas where we can definitely improve the code through error handling, speed ..etc but I figured out this accomplishes the task for now. last note if in prod probably I would consider going with github app sdk as it provides more functionality and easier to implement.

Improvement:
Probably a good idea for to check if response != empty (null) and its array of json incase of changes on the github response or connection errors.

Add unit test
Finally, I tested the code with empty labels and what not and added some handling and the api returns *only* open issues but I put the check anyways in the code.